Edited the configure script

+++ configure.php
@@ -1,4 +1,8 @@
 <?php
 
 // Script that will confugre the widget
-file_put_contents('config.ini', file_get_contents('config.ini.example'));
+
+if (!file_exists('config.ini'))
+{
+  file_put_contents('config.ini', file_get_contents('config.ini.example'));
+}
